refactor: get rid of ui/index file
