+++ src/app/[soorah]/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingBoxes } from '@/ui'
+import { LoadingBoxes } from "@/ui/LoadingBoxes/LoadingBoxes"
 
 export const SoorahLoader = () => <LoadingBoxes />
