+++ src/components/SearchAyah/SearchAyah.tsx
@@ -4,8 +4,9 @@ import { FaExternalLinkAlt } from 'react-icons/fa'
 import Highlighter from 'react-highlight-words'
 
 import { DisplayData } from '@/lib/types'
-import { Sajda } from '@/ui'
+
 import { buildUrl } from '@/utility/buildUrl'
+import { Sajda } from '@/ui/Sajda/Sajda'
 
 export type SearchAyahProps = {
   data: DisplayData
