+++ src/components/SoorahList/SoorahListItem.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 
 import { SoorahListProps } from '@/assets/soorah-list-object'
-import { CityAndSize } from '@/ui'
+import { CityAndSize } from '@/ui/CityAndSize/CityAndSize'
 
 export type SoorahListItemProps = {
   soorah: SoorahListProps
