+++ src/ui/Sajda/Sajda.tsx
@@ -1,6 +1,5 @@
 import { GiArabicDoor } from 'react-icons/gi'
-
-import { Tooltip } from '@/ui'
+import { Tooltip } from '../Tooltip/Tooltip'
 
 export const Sajda = () => (
   <Tooltip message="Səcdə ayəsi">
