+++ src/ui/SoorahCaption/SoorahCaption.tsx
@@ -1,10 +1,10 @@
 import Link from 'next/link'
 
 import { SOORAH_LIST } from '@/assets/soorah-list-object'
-import { CityAndSize } from '@/ui'
 import { buildUrl } from '@/utility/buildUrl'
 
 import styles from './SoorahCaption.module.css'
+import { CityAndSize } from '../CityAndSize/CityAndSize'
 
 type SoorahCaptionProps = {
   soorah: number
