--- src/ui/index.ts
@@ -1,12 +1,0 @@
-export { Bismillah } from './Bismillah/Bismillah'
-export { Card } from './Card/Card'
-export { ColoredText } from './ColoredText/ColoredText'
-export { Footer } from './Footer/Footer'
-export { Loader } from './Loader/Loader'
-export { LoaderProgress } from './LoaderProgress/LoaderProgress'
-export { LoadingBoxes } from './LoadingBoxes/LoadingBoxes'
-export { SoorahCaption } from './SoorahCaption/SoorahCaption'
-export { soorahAyahTitle } from './soorahAyahTitle/soorahAyahTitle'
-export { CityAndSize } from './CityAndSize/CityAndSize'
-export { Tooltip } from './Tooltip/Tooltip'
-export { Sajda } from './Sajda/Sajda'
